pages folder added and loginForm component created
